requirements.txt
AMQPStorm==2.10.7
amqpstorm-flask==0.4.4
aniso8601==9.0.1
attrs==23.2.0
Authlib==1.3.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
cryptography==42.0.8
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
elody[loader]==0.0.86
Flask==3.0.3
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.1.0
inuits-policy-based-auth==9.6.0
inuits-python-logging-loki==1.4.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.0
pamqp==2.3.0
pika==1.3.2
py==1.11.0
py-healthcheck==1.10.1
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.3
pyparsing==3.1.2
python-arango==7.9.1
pytz==2024.1
rabbitmq-pika-flask==1.2.35
rdflib==7.0.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rpds-py==0.18.1
sentry-sdk[flask]==2.4.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
